requirements.txt
pandas==2.2.2
matplotlib==3.8.4
pytest==8.2.0
pytest-timeout==2.3.1
gspread==6.1.2
google-auth==2.31.0
requests==2.32.3
